// src/odsc-vscode-plugin/src/oci-api.d.ts
 export interface IOCIProfile {
     getProfileName(): string;
     getRegionName(): string;
     usesSessionAuth(): boolean;
     getTenancy(): string;
 }
 
 export interface IOCIProfileTreeDataProvider {
     switchProfile(profile: IOCIProfile): Promise<void>;
     refresh(treeItem: IRootNode | undefined): void;
 
     getTreeItem(element: IRootNode): TreeItem;
     getChildren(element?: IRootNode): Thenable<IRootNode[]>;
     findTreeItem(nodeId: string): Promise<IRootNode | undefined>;
 }
 
 export interface IOCIProfileNode extends IRootNode {
     profileName: string;
     regionName: string;
 }
 
 export interface IRootNode {
     updateLabel(newLabel: string): void;
     updateDescription(newDescription: string): void;
     updateParentNode(n: IRootNode): void;
     getParentNode(): IRootNode | undefined;
     getChildren(element: any): Thenable<IRootNode[]>;
 
     id: string;
     label: string;
     commandName: string;
     collapsibleState: TreeItemCollapsibleState;
     lightIcon: string;
     darkIcon: string;
     context: string;
     childrenNodes: IRootNode[];
     parent: IRootNode | undefined;
     description: string | undefined;
     tooltip: string | undefined;
 }
// ...
export class OCIFileExplorerNode extends BaseNode implements IRootNode {
    updateLabel(newLabel: string): void;
    updateDescription(newDescription: string): void;
    updateParentNode(n: IRootNode): void;
    getParentNode(): IRootNode | undefined;
    getChildren(element: any): Thenable<IRootNode[]>;
    label: string;
    commandName: string;
    commandArgs: any[];
    collapsibleState: TreeItemCollapsibleState;
    lightIcon: string;
    darkIcon: string;
    context: string;
    childrenNodes: IRootNode[];
    parent: IRootNode | undefined;
    description: string | undefined;
    tooltip: string | undefined;
    public readonly id: string;
    public readonly uriPath: Uri;
	public readonly type: FileType;
    public canBeDeleted: boolean;
    public isTopDirectory: boolean;
    public isDirectory: boolean;
}